[!/] [Interface] Now we receive timestamps but returns durations..
Before we used one record with (bool duration) for storing timestamps
with signal levels.

Now our interface is that you add timestamps (and signal levels) and
we're returning durations. "Duration" is the same (bool duration) record
but we're storing time difference, not time stamp.

We're added TSignal record for that just to have access to TDuration
field under right name.