[^*] [Implementation] Review and border case fix of internal function..
GetDurationFromMark(Mark) returns part of counter period as duration:

  return Period * (Mark / (MaxMark + 1))

Counter value is TUint_2 but formula needs "+ 1" in divisor.
We want to store that divisor for faster calculations. So we have
to use TUint_4.

Before we stored "divisor + 1" in TUint_2. There were no problems
because our frequency is fixed to 1 KHz and counter's limit for this
freq is less than 65535.